backend/requirements.txt
# FastAPI and core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.13.1

# File system monitoring
watchdog==3.0.0

# Background tasks
celery==5.3.4
redis==5.0.1

# File processing
python-magic-bin==0.4.14
Pillow==10.1.0
PyPDF2==3.0.1

# HTTP client
httpx==0.25.2

# Environment and configuration
python-dotenv==1.0.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
